[spv-out] Move BlockContext into its own file.
Rearrange some code for easier visibility.

`make_local` doesn't need a PhysicalLayout `self`; make it a free function.
